Modified constrained parameters to use bijectors.
The Mapper class now registers a bijection in self.transforms that can
be used to go between the constrained and unconstrained variable.
Assigner now maps assignments on the constrained variables to the
unconstrained variables using the transformers.
A bijector bug made it necessary to implement a temporary fix in
future_features.py.